requirements.txt
arrow==0.8.0
beautifulsoup4==4.5.1
blinker==1.4
cffi==1.8.3
click==6.6
coverage==4.2
cryptography==1.5
cycler==0.10.0
DateTime==4.1.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.4
django-debug-toolbar==1.6
django-disqus==0.5
django-emoticons==1.2
django-forms-bootstrap==3.0.1
django-paypal==0.3.3
django-tinymce==2.4.0
djangorestframework==3.5.2
djangorestframework-jwt==1.8.0
enum34==1.1.6
Flask==0.11.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
gunicorn==19.6.0
html==1.16
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
matplotlib==1.5.3
mock==2.0.0
MySQL-python==1.2.3
numpy==1.11.1
oauthlib==2.0.0
pandas==0.18.1
pandas-datareader==0.2.1
pbr==1.10.0
Pillow==2.9.0
prettytable==0.7.2
pyasn1==0.1.9
pycparser==2.14
pygame==1.9.1
PyJWT==1.4.2
pymongo==3.3.0
pyOpenSSL==16.1.0
pyparsing==2.1.9
pypiwin32==219
python-dateutil==2.5.3
pytz==2016.7
requests==2.11.1
requests-file==1.4
requests-oauthlib==0.6.2
Scrapy==0.16.4
six==1.10.0
SQLAlchemy==1.0.15
sqlparse==0.2.1
static3==0.7.0
stripe==1.38.0
tweepy==3.5.0
Twisted==16.4.1
virtualenv==15.0.3
w3lib==1.15.0
web.py==0.38
Werkzeug==0.11.11
whitenoise==3.2.2
zope.interface==4.3.2
